% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_aniframe.R
\name{aniframe_subclass}
\alias{aniframe_subclass}
\title{The classes an aniframe has in front of \code{aniframe}}
\usage{
aniframe_subclass(data)
}
\arguments{
\item{data}{An aniframe.}
}
\value{
A character vector, e.g. \code{"anipoint"}.
}
\description{
The classes an aniframe has in front of \code{aniframe}
}
\keyword{internal}
