% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aniread-package.R
\docType{package}
\name{aniread-package}
\alias{aniread}
\alias{aniread-package}
\title{aniread: An R package for reading and writing movement data}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

An R package for reading and writing movement data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://animovement.dev/aniread/}
  \item \url{https://github.com/animovement/aniread/}
  \item Report bugs at \url{https://github.com/animovement/aniread/issues}
}

}
\author{
\strong{Maintainer}: Mikkel Roald-Arb<c3><b8>l \email{animovement.84w1m@passmail.com} (\href{https://orcid.org/0000-0002-9998-0058}{ORCID})

Authors:
\itemize{
  \item Mikkel Roald-Arb<c3><b8>l \email{animovement.84w1m@passmail.com} (\href{https://orcid.org/0000-0002-9998-0058}{ORCID})
}

}
\keyword{internal}
