% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_files.R
\name{ensure_file_is_not_empty}
\alias{ensure_file_is_not_empty}
\title{Ensure that the file is not empty.}
\usage{
ensure_file_is_not_empty(path)
}
\arguments{
\item{path}{Path(s) to the file.}
}
\description{
Ensure that the file is not empty.
}
\keyword{internal}
