% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_source.R
\name{has_deeplabcut_csv_header}
\alias{has_deeplabcut_csv_header}
\title{Does a CSV carry a DeepLabCut-style header block?}
\usage{
has_deeplabcut_csv_header(path)
}
\arguments{
\item{path}{Path to the file.}
}
\description{
Does a CSV carry a DeepLabCut-style header block?
}
\keyword{internal}
