% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator_files.R
\name{is_empty_file}
\alias{is_empty_file}
\title{Check if file is empty}
\usage{
is_empty_file(path)
}
\arguments{
\item{path}{Path(s) to the file.}
}
\description{
Check if file is empty.
}
\keyword{internal}
