% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_deeplabcut.R
\name{parse_dlc_pickle}
\alias{parse_dlc_pickle}
\title{Parse DLC pickle string to extract column order}
\usage{
parse_dlc_pickle(pickle_str, multianimal = FALSE)
}
\arguments{
\item{pickle_str}{The values_block_0_kind attribute string}

\item{multianimal}{Whether this is a multi-animal dataset}
}
\value{
A tibble with bodypart, coord columns (and individual if multi-animal)
}
\description{
Parse DLC pickle string to extract column order
}
\keyword{internal}
