% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_source.R
\name{peek_bytes}
\alias{peek_bytes}
\title{Read the leading bytes of a file}
\usage{
peek_bytes(path, n)
}
\arguments{
\item{path}{Path to the file.}

\item{n}{Number of bytes.}
}
\description{
Read the leading bytes of a file
}
\keyword{internal}
