% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_source.R
\name{peek_h5_names}
\alias{peek_h5_names}
\title{Names of the top-level datasets in an HDF5 file}
\usage{
peek_h5_names(path)
}
\arguments{
\item{path}{Path to the file.}
}
\description{
Names of the top-level datasets in an HDF5 file
}
\keyword{internal}
