% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_source.R
\name{peek_header}
\alias{peek_header}
\title{Split the first line of a file into fields}
\usage{
peek_header(path, delim = ",")
}
\arguments{
\item{path}{Path to the file.}

\item{delim}{Field delimiter.}
}
\description{
Split the first line of a file into fields
}
\keyword{internal}
