% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_source.R
\name{peek_lines}
\alias{peek_lines}
\title{Read the first lines of a text file}
\usage{
peek_lines(path, n = 5)
}
\arguments{
\item{path}{Path to the file.}

\item{n}{Number of lines.}
}
\description{
Read the first lines of a text file
}
\keyword{internal}
