% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_deeplabcut.R
\name{read_deeplabcut_csv_multi}
\alias{read_deeplabcut_csv_multi}
\title{Read multi-animal DLC files}
\usage{
read_deeplabcut_csv_multi(path)
}
\description{
Read multi-animal DLC files
}
\keyword{internal}
