% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_deeplabcut.R
\name{read_deeplabcut_csv_single}
\alias{read_deeplabcut_csv_single}
\title{Read single-animal DLC files}
\usage{
read_deeplabcut_csv_single(path)
}
\description{
Read single-animal DLC files
}
\keyword{internal}
