% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_registry.R
\name{registry_sources}
\alias{registry_sources}
\title{Names of the sources that can be read or written}
\usage{
registry_sources(direction)
}
\arguments{
\item{direction}{\code{"read"} or \code{"write"}.}
}
\value{
A character vector of source names, in registry order.
}
\description{
Names of the sources that can be read or written
}
\keyword{internal}
