% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_octron.R
\name{resolve_area_sum}
\alias{resolve_area_sum}
\title{Sum all segment areas per row (used for \code{area} under \code{method = "weighted"}).}
\usage{
resolve_area_sum(areas_list)
}
\description{
Sum all segment areas per row (used for \code{area} under \code{method = "weighted"}).
}
\keyword{internal}
