% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trackball.R
\name{resolve_column}
\alias{resolve_column}
\title{Resolve column identifier to column name}
\usage{
resolve_column(data, col)
}
\arguments{
\item{data}{Data frame}

\item{col}{Column identifier (name or index)}
}
\value{
Column name as character
}
\description{
Resolve column identifier to column name
}
\keyword{internal}
