% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_source.R
\name{run_detector}
\alias{run_detector}
\title{Run one detector, treating any failure as a non-match}
\usage{
run_detector(detector, path)
}
\arguments{
\item{detector}{A detector function.}

\item{path}{Path to the file.}
}
\description{
Run one detector, treating any failure as a non-match
}
\keyword{internal}
