% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aniread-package.R
\docType{package}
\name{aniread-package}
\alias{aniread}
\alias{aniread-package}
\title{aniread: An R toolbox for reading and writing movement data}
\description{
An R toolbox for reading and writing movement data.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.roald-arboel.com/aniread/}
  \item \url{https://github.com/roaldarbol/aniread/}
  \item Report bugs at \url{https://github.com/roaldarbol/aniread/issues}
}

}
\author{
\strong{Maintainer}: Mikkel Roald-Arbøl \email{aniread.fi28v@passmail.com} (\href{https://orcid.org/0000-0002-9998-0058}{ORCID})

}
\keyword{internal}
