% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_octron.R
\name{detect_octron_tuple_cols}
\alias{detect_octron_tuple_cols}
\title{Detect columns that contain tuple-strings (e.g. "(120.5, 85.3)")}
\usage{
detect_octron_tuple_cols(data)
}
\description{
Detect columns that contain tuple-strings (e.g. "(120.5, 85.3)")
}
\keyword{internal}
