% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_octron.R
\name{parse_octron_column}
\alias{parse_octron_column}
\title{Parse a column of mixed scalars and tuple-strings into a list of
numeric vectors (one per row).}
\usage{
parse_octron_column(col)
}
\description{
Parse a column of mixed scalars and tuple-strings into a list of
numeric vectors (one per row).
}
\keyword{internal}
