% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_custom.R
\name{read_custom}
\alias{read_custom}
\title{Read a custom file format}
\usage{
read_custom(path, cols, metadata = list())
}
\arguments{
\item{path}{Character string specifying the path to the file to read.}

\item{cols}{A named vector specifying which columns to keep and how to rename
them, e.g. \code{c(time = "time", x = "x", y = "y")}. Names should be the desired
output column names, and values can be either the original column names (as
characters) or column positions (as integers, e.g. \code{c(time = 1, x = 2, y = 3)}).
.}

\item{metadata}{A list of metadata to attach to the resulting aniframe.
Default is an empty list. To see which metadata fields can be set, see aniframe::default_metadata().}
}
\value{
An aniframe with the selected and renamed columns, and attached
metadata.
}
\description{
Reads a file and selects/renames specified columns to create an aniframe.
}
\examples{
\dontrun{
# Using column names
read_custom("data.csv", cols = c(time = "frame", x = "pos_x", y = "pos_y"))

# Using column positions
read_custom("data.csv", cols = c(time = 1, x = 2, y = 3))
}
}
