% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_deeplabcut.R
\name{read_deeplabcut}
\alias{read_deeplabcut}
\title{Read DeepLabCut data}
\usage{
read_deeplabcut(path)
}
\arguments{
\item{path}{Path to a DeepLabCut data file}
}
\value{
an aniframe
}
\description{
Read files from DeepLabCut (DLC) in either csv or h5 format.
}
