% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_deeplabcut.R
\name{read_deeplabcut_csv}
\alias{read_deeplabcut_csv}
\title{Read DeepLabCut data}
\usage{
read_deeplabcut_csv(path, multianimal = NULL)
}
\arguments{
\item{path}{Path to a DeepLabCut data file}

\item{multianimal}{By default, whether a file is multi-animal is detected automatically. This gives an option to ensure it. logical TRUE/FALSE.}
}
\value{
a movement dataframe
}
\description{
Read csv files from DeepLabCut (DLC). The function recognises whether it is a
single- or multi-animal dataset.
}
\keyword{internal}
