% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_deeplabcut.R
\name{read_deeplabcut_h5}
\alias{read_deeplabcut_h5}
\title{Read DeepLabCut H5 file}
\usage{
read_deeplabcut_h5(path)
}
\arguments{
\item{path}{Path to the DLC .h5 file}
}
\value{
An aniframe with columns: time, individual, keypoint, x, y, confidence
}
\description{
Read DeepLabCut H5 file
}
\keyword{internal}
