% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasttrack.R
\name{read_fasttrack}
\alias{read_fasttrack}
\title{Read FastTrack tracking data}
\usage{
read_fasttrack(path)
}
\arguments{
\item{path}{Path to a FastTrack tracking.txt file.}
}
\value{
An aniframe
}
\description{
Reads a FastTrack tracking result file (.txt format) and returns an
aniframe with keypoints for head, body, and tail positions.
}
