% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lightningpose.R
\name{read_lightningpose}
\alias{read_lightningpose}
\title{Read LightningPose data}
\usage{
read_lightningpose(path)
}
\arguments{
\item{path}{Path to a LightningPose data file}
}
\value{
an aniframe
}
\description{
Read csv files from LightningPose (LP).
}
