% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trackball.R
\name{read_trackball}
\alias{read_trackball}
\title{Read trackball data}
\usage{
read_trackball(
  paths,
  setup = c("of_free", "of_fixed"),
  sampling_rate,
  col_time = "time",
  col_dx = "x",
  col_dy = "y",
  counts_per_rotation = NULL,
  ball_diameter = NULL,
  dots_per_cm = NULL,
  quiet = TRUE
)
}
\arguments{
\item{paths}{Two file paths, one for each sensor (although one is allowed for a fixed setup, \code{of_fixed}).}

\item{setup}{Which type of experimental setup was used. Expects either \code{of_free} or \code{of_fixed}.}

\item{sampling_rate}{Sampling rate tells the function how long time it should integrate over. A sampling rate of 60(Hz) will mean windows of 1/60 sec are used to integrate over.}

\item{col_time}{Which column contains the information about time. Can be specified either by the column number (numeric) or the name of the column if it has one (character). Should either be a datetime (POSIXt) or seconds (numeric).}

\item{col_dx}{Column name for x-axis values}

\item{col_dy}{Column name for y-axis values}

\item{counts_per_rotation}{For \code{of_fixed} setup: the sensor count for a full 360 degree rotation. Can be obtained using \code{calibrate_trackball()}.}

\item{ball_diameter}{For \code{of_fixed} setup: the ball diameter (in same units as desired output). Required if using \code{dots_per_cm} instead of \code{counts_per_rotation}.}

\item{dots_per_cm}{For \code{of_fixed} setup: sensor dots-per-cm. Use with \code{ball_diameter} as an alternative to \code{counts_per_rotation}.}

\item{quiet}{If \code{TRUE} (default), suppresses most warning messages.}
}
\value{
a movement dataframe
}
\description{
Read trackball data from a variety of setups and configurations.
}
