% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_octron.R
\name{resolve_largest}
\alias{resolve_largest}
\title{Pick the value of the segment with the largest area per row.}
\usage{
resolve_largest(values_list, areas_list)
}
\description{
Pick the value of the segment with the largest area per row.
}
\keyword{internal}
